backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.6
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.14
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
redis==4.1.1
requests==2.27.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.0.1
ua-parser==0.10.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.6.0
